l'ajout de la couche metier

+++ src/main/java/com/enset/tp/hospitalmanagement/repositories/RendezVousRepository.java
@@ -3,5 +3,9 @@ package com.enset.tp.hospitalmanagement.repositories;
 import com.enset.tp.hospitalmanagement.entities.RendezVous;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RendezVousRepository extends JpaRepository<RendezVous, Integer> {
+import java.util.Optional;
+import java.util.concurrent.Future;
+
+public interface RendezVousRepository extends JpaRepository<RendezVous, Long> {
 }
+
